ha-moode-radios-addon/app/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
pydantic==2.11.4
pydantic-settings==2.9.1
Pillow==11.2.1
asyncssh==2.21.0
